Type Model (#4)
This PR models almost all of the types that will be necessary for
optimization. This includes:

- generic relational algebra operators that allow us to use the same
"type" for both expressions in the memo table and operators in the plans
- logical / physical plans
- scalar operators and expressions
- partially materialized logical plans for rule binding
- transformation rule + implementation rule trait and some empty structs
that implement them

I've named the crate itself `optd-core`. This can be subject to change,
but I feel this is a reasonable default for now.

~~TODO: need to wait on #3 and #12 to be merged before proper CI checks
can happen~~

Edit: I removed the `cargo rustdoc` check because its creating more
problem than it would solve, see #14

---------

Co-authored-by: Alexis Schlomer <aschlome@andrew.cmu.edu>