First PR cleanup (#18)
## Problem

#4 Was the first major PR, but since we were making a lot of changes
last minute there were a few things that slipped through with respect to
quality control.

## Summary of changes

- Added very strict clippy documentation requirements. Discussion below.
- Added documentation for everything except the
`operator::relational::physical` module and the `operator::scalar`
module, as those seem kind of unstable right now.
- Changed the scalar generic param of expressions to use `ScalarGroupId`
instead of `GroupId`.
- Added several TODOs that I think will be resolved pretty quickly,
though I'd appreciate it if others took a look at those in the case that
we can actually get those done now.

Now that we have something to work off of, requiring documentation on
everything (including private items) means that we'll pay the cost right
now of making sure people understand what we're doing in the future. I'm
willing to relax it a little bit if we put _other_ stuff in place that
ensure nobody in the future encounters large chunks of code with zero
documentation.