feat: DSL parser rewrite (#25)
## Problem
The previous parser was imperative, had poorly defined syntax, was
missing the entire physical side, and was not maintainable (essentially
a hackathon-like POC). This PR solves all of these problems by designing
a better language and using a declarative parser.

## Summary of changes
- Better, more simplistic AST
- Support for ADTs
- Parsing with error recovery and useful messages
- Support map literals
- Much, much more expressive syntax to model query optimization (see
test cases for now, will release a formal specification eventually)