feat: stabilize memo goals (#33)
## Problem

Due to the need to handle physical property, the physical expression is
now organized to be memoized based on an optimization _goal_, which
represents a (group, required_physical_properties) pair instead of a
group alone.

Admittedly, we could share expressions among the different goals for the
same group. However, it is hard to consider sharing expressions for all
potential goals (a different combination of properties). It is likely in
practice we will share an expression only with the default goal (with no
required properties).

## Summary of changes


**Memo**
- stabilize Memo API `create_or_get_goal`, `add_physical_expr_to_goal`,
`get_goal`, and `get_winner_physical_expr_in_goal`.

**Storage**
- A new `goals` table that mimics the `relation_groups` table for
physical optimizations.

_Misc_

-  fix some clippy lints for `optd-dsl`

---------

Signed-off-by: Yuchen Liang <yuchenl3@andrew.cmu.edu>