Clean up doc comments and add periods (#43)
## Problem

Cleans up a bit of the CPS PR.

## Summary of changes

This commit adds periods to all of the comments in the cir and engine. I
also added some more detail here and there.

This commit also unifies all of the types in the cir to make it easier
to interact with those types. Imo it makes more sense to have them all
in one place rather than to have to figure out the paths to each of them
every time you import.