Refactor rule engine + migrate to `optd-dsl` (#47)
## Problem

We want to have the `engine` module located inside the `optd-dsl` crate
so that we don't end up with a circular dependency when trying to allow
the dsl "UDF"s to access the `Engine`.

```rust
// Right now (defined in `optd-dsl`):
RustUDF(fn(Vec<Value>) -> Value),

// We want something similar to:
RustUDF(fn(Engine, Vec<Value>) -> Value)
//         ^^^^^^ If this was in the `optd-core` crate we would have a circular dependency
```

## Summary of changes

There was a lot of duplicate code and unnecessary function signatures,
so I cleaned all of it up.

Additionally, the entire `engine` module was moved into `optd-dsl`. This
caused a lot of problems, so a lot of function signatures had to be
changed.

Notably, I got rid of the `Evaluate` trait inside `optd-core` and added
an `evaluate` method directly on `Engine`. I still think that is would
be good to have every `evaluate_???` function be a private method on
`Engine` for readability purposes, but that can be a different PR.

(It might be good to look at the individual commits rather than the
entire diff)

---------

Co-authored-by: Alexis Schlomer <aschlome@andrew.cmu.edu>