feat: Generic CPS (#48)
## Problem

We need a way to finish a job on a tokio task and use the correct job ID
to send a message back to the main event loop. Before this PR, we always
returned the unit type for continuations. We achieve correctness while
maintaining good abstractions between the engine and the optimizer by
making continuation return an `EngineResponse<O>` with a converter
function into some output type.

```rust
/// The engine response type, which can be either a return value with a converter callback
/// or a yield group/goal with a continuation for further processing.
pub enum EngineResponse<O> {
    /// The engine has returned a value, and the final continuation
    /// should be called on it to produce the desired output.
    Return(Value, Continuation<Value, O>),
    /// The engine has yielded a group ID, and the continuation
    /// should be called for each value after expanding the group.
    YieldGroup(GroupId, Continuation<Value, EngineResponse<O>>),
    /// The engine has yielded a goal, and the continuation
    /// should be called for each value after expanding the goal.
    YieldGoal(Goal, Continuation<Value, EngineResponse<O>>),
}
```

## Summary of changes

- Remove the generator, we don't need it anymore.
- No longer needs to pass channels to the engine.
- We can construct the `OptimizerMessage` using the final callback, but
send messages outside the chain of continuations.
- Keeping the initial callback (or we call it the `transform`
continuation):
- We use it for validating the output `Value` to CIR representation and
constructing the single message type.
- The output type is made generic so `optd-dsl` does not have recursive
dependencies.

### Notes for reviewers

New signatures for the future and continuation:


https://github.com/cmu-db/optd/blob/272fc6842213c7766f9824f5dbd298fa8b19ff2b/optd-dsl/src/engine/utils.rs#L10-L17

The only place we return a `YieldGroup`:


https://github.com/cmu-db/optd/blob/272fc6842213c7766f9824f5dbd298fa8b19ff2b/optd-dsl/src/engine/eval/match.rs#L182-L189

The only place we return a `YieldGoal`:


https://github.com/cmu-db/optd/blob/272fc6842213c7766f9824f5dbd298fa8b19ff2b/optd-dsl/src/engine/eval/match.rs#L193-L199



**Note:** The state of the optimizer on `main` is very behind what we
have right now so I commented a lot of stuff out, but I made all engine
tests pass and thought about the cases for integrating it with the
optimizer.

---------

Signed-off-by: Yuchen Liang <yuchenl3@andrew.cmu.edu>
Co-authored-by: Alexis Schlomer <aschlome@andrew.cmu.edu>