(feat) AST->HIR Code prep (#53)
This PR solves a bunch of problems in the DSL/HIR to allow for the
initial transformation from AST -> HIR. Notably:

1. Adds  `$` and `*` type extensions in the parser & type system.

`Type$` stands for Costed(Type) is covariant and inherits Stored(Type)
`Type* `stands for Stored(Type) is covariant and inherits Type.

2. Adds an implementation of map in the HIR and the engine where the
keys are allowed to be hashed and compared against each other.

3. Adds `#` to distinguish between field accesses in structs, and member
function calls (with `.`).

4. Fixes the `Call` evaluate engine function to include indexing arrays,
tuples, maps, and operators (including a group/goal expansion when field
access is needed).

5. Added the `Never` type for the Fail expression in the type system.