(feat) Semantic Check: ADTs et alia (#60)
### Semantic Analysis

Performs the following semantic analysis for our DSL.

### adt_check

In `semantic_check/adt_check.rs`:

1. Detects cyclic ADT definitions. This is best described with some
pictures:


![image](https://github.com/user-attachments/assets/610597c9-3493-4b08-bfda-cd0e110662cb)

or, more elaborate:


![image](https://github.com/user-attachments/assets/54b02716-723f-46e0-9545-a0f6695c64bb)

2. Detects whether `Logical`, `Physical`, `LogicalProperties`, and
`PhysicalProperties` types are defined, and are not the children of any
other types.

3. Checks for duplicate fields in ADTs.


![image](https://github.com/user-attachments/assets/afe3683c-9283-4d73-a468-11c58ba89e3b)

4. Validates the types within the fields of ADTs. Specifically:

-  No costed or stored types.
- Only logical types may contain logical children.
- Only physical types may contain physical children.
- Logical `product` types may only contain [Logical] or Logical fields.
- Physical `product` types may only contain [Physical] or Physical
fields.

### from_ast

In `from_ast`, I made the code process ADT definition first. This allows
us to perform the following extra semantic analysis:

1. Verify if `Constructors` and `Patterns` are properly defined.


![image](https://github.com/user-attachments/assets/19926c84-eea5-47e9-b5a5-12c30994b36b)

2. Verify if all annotated types exist when converting them into their
`HIR` version.

### bug fix & others

- The parser was parsing `None` as a `Constructor`. I have changed the
syntax to `none` lowercase, and fixed the parser code correspondingly.

- Refactored `AnalyzerErrorKind` to be `Boxed`, as clippy was
complaining the type was getting too big.

- Made `subtypes` in `TypeRegistry` a `BTreeMap` so that iterating over
all keys is deterministic across compilations.

- One semantic check I didn't implement is verifying the number of
arguments in constructors, patterns, and function calls. **This will be
addressed in a future PR for patterns and constructors**. For function
calls, this validation will be handled in the type inference PR since
functions are first-class values in our language, making it more
sensible to perform there (especially since calls are also used for
arrays, maps, and tuples).