Add top-level `optd` crate + Fix CI + Fix `Cargo.toml`s (#66)
Eventually, we will want `optd` to be an optimizer service. However, we
would also like it to be an embeddable library. This PR creates a
top-level `optd` crate that will act as the "glue" between all of our
components (which right now is just `optd-core` and `optd-dsl`, but more
will probably be created). Users should be able to import a single
`optd` dependency and have access to all functionality we want to
provide (similar to `tokio`). We do _not_ want a datafusion situation
where there could be 5 different datafusion crates you need to import.

There is another subtle reason we need this now. We would like to add
catalog functionality to `optd`. Both the rule engine and the optimizer
need access to the catalog. However, it is not entirely obvious where
one would put an implementation of a catalog right now. It doesn't
really make sense to put it in `optd-dsl`, but we also can't put it in
`optd-core` because `optd-core` depends on `optd-dsl`, which would
result in a circular dependency.

By creating a top-level `optd` crate, we can put the definition of a
`Catalog` trait in `optd-dsl`, and then add implementations of catalogs
in the `optd` crate. This allows for flexibility in catalog
implementations while not adding (arguably) irrelevant code to the
`optd-dsl` crate.

I removed some unused dependencies and bumped all `Cargo.toml`
dependencies to their latest version. However, I updated the
`Cargo.lock` with `-Z direct-minimal-versions`.

This PR also reintroduces the CI checks that we disabled. The
`twox-hash` problem still exists when taking a dependency on `iceberg`,
so instead of `-Z minimal-versions`, I've changed it to `-Z
direct-minimal-versions`.