Create `compile_hir` compilation entrypoint. (#77)
## Problem

The only place one could compile an HIR was through the `optd-cli`
binary. This PR brings this behavior back into the library so we can
extend its functionality with UDFs (in a future PR).

## Summary of changes

Created a `Config` struct inside `compile.rs` that largely mimics the
previous `clap` struct but cleans it up a bit. This `Config` contains a
flattened `Verbosity` struct that groups the different verbosity
settings together when compiling. This makes it much easier to pass all
of the configurations directly to functions without expanding the
fields.

The `compile_hir` function is also located in `compile.rs`, and it
largely copies the function that was previously in
`optd/src/dsl/cli/main.rs`. This will make it easy to modify what we can
pass into compilation (namely a function table for UDFs).