Implement UDF linking (#78)
## Summary of changes

This PR implements linking of UDFs by passing in a table of function
pointers into `compile_hir`.

This PR also adds the `Catalog` field to both the rule engine and the
optimizer as a trait object (`Arc<dyn Catalog>`). I strongly considered
making this a generic parameter, but the problem with that is then the
`C: Catalog` type parameter would need to be added to literally every
structures and function, and I that would significantly reduce
readability. The performance is essentially negligible because it is
unlikely the compiler could make significant optimizations based on the
implementation of the `Catalog`.

Ideally I would add tests that test the functionality of the catalog as
well but I don't have time (and also I don't have any reason to think
that it wouldn't work now). Once I get an interface that the optimizer
needs from the catalog from @SarveshOO7 and @yliang412, I can add those
tests easily.