(feat) HIR Final Compilation Pass  (#87)
### About

This PR performs the final transformation pass from the `HIR<TypedSpan>`
to the `HIR`.
**It is now possible to run rules in the optimizer using the DSL.**
:partying_face: :tada:

### Changes

I added a new `into_hir` module, similar to the `from_ast` module, which
I refactored the interface for consistence.
The conversion essentially performs three things:

1. Removing type annotations from expressions and patterns
2. Converting struct expressions to operator expressions where
appropriate
3. Transforming field access expressions into indexing operations

### Future changes

Validating the annotations (e.g. transformation, implementation, UDF) is
the next planned PR. It should also be done in that pass.