Associated Error Types (#104)
## Problem

Because the memo table implementations are likely going to have very
different error conditions, we would ideally like the implementors of
the memo table traits to define their errors instead of us.

For example, a persistent memo table on disk will want to return I/O
errors, but an in-memory memo table should be infallible. A memo table
as a service over the network might have completely different errors,
and a parallel memo table could have different behavior as well.

## Summary of changes

This PR modifies the `memo` module to use a new `MemoBase` trait that
allows the implementors to define an associated error type.

```rust
/// Base trait defining a shared implemention-defined error type for all memo-related traits.
pub trait MemoBase {
    /// The error type used throughout all memo-related traits.
    type MemoError: Debug;
}
```

The reason we need this is because the memo traits have been effectively
divided into 3 subtraits, and we cannot specify 1 associated type for 3
different traits.

---

Note that the error type for the in-memory memo table is `Infallible`.

```rust
/// The never type. Used for ensuring that the in-memory memo table never raises an error.
#[derive(Debug)]
pub enum Infallible {}

/// This defines the error type for the in-memory implementation of the memo table.
impl MemoBase for MemoryMemo {
    /// In-memory operations cannot fail.
    type MemoError = Infallible;
}
```

This could change in the future, but at least for now it is not really
clear what a user of the in-memory table would do if it encountered a
broken invariant (a group does not exist, for example). So this PR
simply makes all of those error cases a panic / crash. This can easily
be changed in the future.

# TODO

- [x] Figure out how to error handle in the optimizer code
- [x] Refactor how the `optimizer` module handles errors with
`OptimizerError`, since `MemoError` used to be the only variant
- [x] Clean up lints