Some bug fixes in logical -> logical optimizer (#109)
The logical -> logical optimizer was untested, this PR adds demo/test
code for sanity checking during development. It ensures that the
optimizer does not crash, and that the memo table gets populated with
expressions (through prints).

This demo code will go away as we get a more robust testing framework &
observation of the task graph.